refactor: feed 생성 css 변경
