--- src/styles/components/feed/_feed-create-description.scss
@@ -1,17 +1,0 @@
-.feed-create-description__textarea {
-  width: 100%;
-  height: 8rem;
-  padding: 0.25rem;
-  font-size: var(--font-size-lg);
-  color: var(--white);
-  background-color: transparent;
-  resize: none;
-  border: 1px solid var(--white);
-}
-
-.feed-create-form-description__image {
-  padding: 3rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
