+++ src/styles/components/feed/_feed-create.scss
@@ -9,58 +9,80 @@
   height: 100%;
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
 }
 
 .feed-create-form-image {
+  position: relative;
   display: flex;
   flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  color: var(--white);
-  font-size: large;
-}
-
-.feed-create-form-image__button {
-  width: 3rem;
-}
-
-.feed-create-form-image-blank {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
 }
 
 .feed-create-form-image__button--delete {
-  display: flex;
-  justify-content: flex-end;
+  position: absolute;
+  right: 0.5rem;
+  z-index: 10;
 }
 
 .feed-create-form-image-upload {
   display: flex;
-  flex-direction: column;
-  justify-content: center;
+  flex-direction: row;
+  margin-bottom: 5rem;
+  overflow: scroll;
+  -ms-overflow-style: none; /* IE and Edge */
+  scrollbar-width: none; /* Firefox */
+}
+
+.feed-create-form-image-upload::-webkit-scrollbar {
+  display: none; /* Chrome, Safari, Opera*/
 }
 
 .feed-create-form-image-upload__image {
-  img {
-    object-fit: scale-down;
-  }
-}
+  display: flex;
+  border-radius: 10%;
+  margin-right: 0.5rem;
 
-.feed-create-form-image-text {
-  margin-top: 1rem;
-  color: var(--white);
+  img {
+    object-fit: fill;
+    border-radius: 10%;
+  }
 }
 
 .feed-create-form-input {
   flex-grow: 0;
-  margin-bottom: 3rem;
+  margin: 5rem 0rem;
   &__label {
+    display: flex;
+    justify-content: center;
     cursor: pointer;
   }
 
   &__input {
     display: none;
   }
+}
+
+.feed-create-form-input__button-icon {
+  margin-right: 1rem;
+}
+
+.feed-create-form-content {
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.feed-create-form-complete__button {
+  margin-bottom: 1.5rem;
+}
+
+.feed-create-form__textarea {
+  width: 100%;
+  height: 8rem;
+  padding: 0.25rem;
+  font-size: var(--font-size-lg);
+  color: var(--white);
+  background-color: transparent;
+  resize: none;
+  border: 1px solid var(--white);
 }
