+++ src/styles/components/feed/_index.scss
@@ -1,6 +1,5 @@
 @use './feed-create';
 @use './feed-create-header';
-@use './feed-create-description';
 @use '../feed/[id]' as c;
 
 @import '../../utils/';
